requirements.txt
fastapi<0.100.0
pydantic<2.0.0
uvicorn
transformers
torch
pillow
python-multipart
openai
